Move stuff from cra to vite
