+++ frontend/.env.example
@@ -1,1 +1,1 @@
-REACT_APP_API_URL=http://localhost:3000/api
+VITE_API_URL=http://localhost:3000/api
