+++ frontend/package.json
@@ -17,36 +17,19 @@
     "react-lazy-load": "^4.0.1",
     "react-redux": "^8.0.4",
     "react-router-dom": "^6.26.1",
-    "react-scripts": "5.0.1",
     "react-select": "^5.8.0",
     "react-tooltip": "^5.28.0",
     "recharts": "^2.12.7",
     "redux": "^4.2.1",
-    "universal-cookie": "^4.0.4",
-    "web-vitals": "^2.1.4"
+    "universal-cookie": "^4.0.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.7.0",
+    "vite": "^6.3.5"
   }
 }
