+++ frontend/src/components/Legal.jsx
