+++ frontend/src/components/StatsData.jsx
