+++ frontend/src/components/games/components/ChampionDetails.jsx
