+++ frontend/src/components/games/components/GameTitle.jsx
