+++ frontend/src/components/games/splashArtGame.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useCallback } from "react";
 import axios from "axios";
 import Select from "react-select";
-import Victory from "./components/victory";
-import ItemImg from "./components/itemImg";
+import Victory from "./components/Victory";
+import ChampionImg from "./components/ChampionImg";
 import Config from "../../configs/config";
 import {
   saveGamesPlayed,
@@ -10,25 +10,27 @@ import {
   saveFirstTries,
 } from "../../utils/saveStats";
 import { Reroll } from "../../utils/reroll";
+import LazyLoad from "react-lazy-load";
 import {
-  customFilterOptionItems,
-  HoverSelectStyles,
+  SelectStyles,
+  customFilterOptionChamps,
   SelectTheme,
 } from "./styles/selectStyles";
 import { useSelector } from "react-redux";
 import {
-  getOldItemGuessHistory,
-  addToOldItemGuessHistory,
-  clearOldItemHistory,
+  clearSkinHistory,
+  getSkinGuessHistory,
+  addToSkinGuessHistory,
 } from "../history";
 
-export default function OldItemGame() {
+export default function SplashArtGame() {
   const [validGuesses, setValidGuesses] = useState([]);
-  const [items, setItems] = useState([]);
+  const [champions, setChampions] = useState([]);
   const [guesses, setGuesses] = useState([]);
-  const [currentGuess, setGuess] = useState();
+  const [currentGuess, setGuess] = useState(validGuesses[0]);
   const [correctGuess, setCorrectGuess] = useState(false);
   const [sprite, setSprite] = useState("");
+  const [title, setTitle] = useState("");
 
   const isColorBlindMode = useSelector(
     (state) => state.colorBlindReducer.isColorBlindMode
@@ -43,40 +45,39 @@ export default function OldItemGame() {
   );
 
   useEffect(() => {
-    FetchItems();
-    FetchItemImage();
+    FetchChampions();
+    FetchSplashArt();
     SetHistory();
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   const SetHistory = () => {
-    const history = getOldItemGuessHistory().reverse();
+    const history = getSkinGuessHistory().reverse();
 
     if (history.length > 0) {
-      setItems(history);
+      setChampions(history);
       setGuesses(history.map((item) => item.name));
     }
   };
 
-  const FetchItems = () => {
+  const FetchChampions = () => {
     axios
-      .get(Config.url + "/oldItems")
+      .get(Config.url + "/champions")
       .then((response) => {
         if (response.data.status === "success") {
-          const data = response.data.items;
+          const data = response.data.champions;
           data.sort((a, b) => a.value.localeCompare(b.value));
 
-          // Remove already guessed items from valid guesses
-          const guessHistoryNames = new Set(
-            getOldItemGuessHistory().map((item) => item.name)
+          const guessChampionKeys = new Set(
+            getSkinGuessHistory().map((champ) => champ.key)
           );
 
           const transformedData = data
-            .filter((item) => !guessHistoryNames.has(item.value))
+            .filter((champion) => !guessChampionKeys.has(champion.value))
             .map((champion) => ({
               value: champion.value,
               label: champion.value,
+              image: champion.image,
             }));
-
           setValidGuesses(transformedData);
         }
       })
@@ -85,34 +86,9 @@ export default function OldItemGame() {
       });
   };
 
-  const ApplyBlur = useCallback(
-    (guessCount) => {
-      const spriteImg = document.getElementById("spriteImg");
-      if (!spriteImg) return;
-
-      const initialBlur = 1.0;
-      let blurVal = initialBlur;
-
-      for (let i = 0; i < guessCount; i++) {
-        blurVal -= blurVal * 0.4;
-      }
-
-      spriteImg.style.filter = `blur(${blurVal.toFixed(3)}em) ${
-        isMonochrome ? "grayscale(1)" : ""
-      }`;
-    },
-    [isMonochrome]
-  );
-
-  useEffect(() => {
-    if (sprite) {
-      ApplyBlur(guesses.length);
-    }
-  }, [sprite, guesses, isMonochrome, ApplyBlur]);
-
-  const FetchItemImage = () => {
+  const FetchSplashArt = () => {
     axios
-      .get(Config.url + "/oldItem", {
+      .get(Config.url + "/splash", {
         headers: { authorization: "Bearer " + localStorage.getItem("token") },
       })
       .then((response) => {
@@ -143,7 +119,7 @@ export default function OldItemGame() {
 
     axios
       .post(
-        Config.url + "/oldItem",
+        Config.url + "/splash",
         { guess: currentGuess },
         {
           headers: { authorization: "Bearer " + localStorage.getItem("token") },
@@ -153,16 +129,16 @@ export default function OldItemGame() {
         if (response.data.status !== "success") {
           return;
         }
-
         saveTries(1);
 
         const isCorrect = response.data.correctGuess;
-        const itemId = response.data.itemId;
+        const key = response.data.championKey;
+
         const name = response.data.name;
 
-        setItems((items) => [{ id: itemId, name, isCorrect }, ...items]);
-
-        addToOldItemGuessHistory({ id: itemId, name, isCorrect });
+        // Use object instead of tuple
+        setChampions((champions) => [{ key, isCorrect, name }, ...champions]);
+        addToSkinGuessHistory({ key, isCorrect, name });
 
         const spriteImg = document.getElementById("spriteImg");
 
@@ -172,54 +148,81 @@ export default function OldItemGame() {
           }
           saveGamesPlayed();
           setCorrectGuess(true);
+          clearSkinHistory();
+          setTitle(response.data.title);
+
           spriteImg.style.filter = "";
-          clearOldItemHistory();
         } else {
           ApplyBlur(guesses.length + 1);
         }
       })
       .catch((error) => {
         console.log(error);
-        setItems([]);
+        setChampions([]);
       });
   };
+
+  const ApplyBlur = useCallback(
+    (guessCount) => {
+      const spriteImg = document.getElementById("spriteImg");
+      if (!spriteImg) return;
+
+      const initialBlur = 1.0;
+      let blurVal = initialBlur;
+
+      for (let i = 0; i < guessCount; i++) {
+        blurVal -= blurVal * 0.4;
+      }
+
+      spriteImg.style.filter = `blur(${blurVal.toFixed(3)}em) ${
+        isMonochrome ? "grayscale(1)" : ""
+      }`;
+    },
+    [isMonochrome]
+  );
+
+  useEffect(() => {
+    if (sprite) {
+      ApplyBlur(guesses.length);
+    }
+  }, [sprite, guesses, isMonochrome, ApplyBlur]);
 
   const Restart = () => {
     setTimeout(() => ApplyBlur(0), 0);
 
-    FetchItemImage();
-    FetchItems();
+    FetchSplashArt();
+    FetchChampions();
 
     setGuesses([]);
-    setItems([]);
+    setChampions([]);
     setGuess();
     setCorrectGuess(false);
-    clearOldItemHistory();
   };
 
   const HandleReroll = () => {
-    clearOldItemHistory();
+    clearSkinHistory();
     setGuesses([]);
-    setItems([]);
-    Reroll("oldItem");
+    setChampions([]);
+    Reroll("splash");
   };
 
   return (
     <div className="container main pt-4 pb-5 mb-5">
-      <h3 className="text-center pb-3">Which legacy item is this?</h3>
+      <h3 className="text-center pb-3">Whose splash art is this?</h3>
 
       <div
         className="container d-flex justify-content-center shadow"
-        id="itemContainer"
+        id="spriteContainer"
       >
         <img
           src={`data:image/webp;base64,${sprite}`}
           style={{
+            filter: `${isMonochrome ? "grayscale(1)" : ""}`,
             transform: `${randomRotate ? "rotate(180deg)" : ""}`,
           }}
           className="rounded p-4"
           id="spriteImg"
-          alt="Item sprite."
+          alt="Champion splash art."
           draggable="false"
         />
       </div>
@@ -231,18 +234,25 @@ export default function OldItemGame() {
           id="guess-form"
         >
           <Select
+            className="select"
             options={validGuesses}
             onChange={(selectedOption) => setGuess(selectedOption.value)}
             isDisabled={correctGuess}
-            placeholder="Type items name"
-            filterOption={customFilterOptionItems}
-            styles={HoverSelectStyles}
-            theme={SelectTheme}
+            styles={SelectStyles}
+            placeholder="Type champions name"
+            filterOption={customFilterOptionChamps}
             formatOptionLabel={(data) => (
               <div className="select-option">
+                <LazyLoad offset={200}>
+                  <img
+                    src={"/40_40/champions/" + data.image + ".webp"}
+                    alt="Champion icon"
+                  />
+                </LazyLoad>
                 <span>{data.label}</span>
               </div>
             )}
+            theme={SelectTheme}
           />
 
           <div className="d-flex justify-content-evenly">
@@ -273,14 +283,13 @@ export default function OldItemGame() {
       </div>
 
       <div id="championsImgs" className="container">
-        {items.map((item) => (
-          <ItemImg
-            key={item.id}
-            itemId={item.id}
-            name={item.name}
-            isCorrect={item.isCorrect}
-            path="/old_items/"
+        {champions.map((champ) => (
+          <ChampionImg
+            key={champ.key}
+            championKey={champ.key}
+            isCorrect={champ.isCorrect}
             isColorBlindMode={isColorBlindMode}
+            name={champ.name}
           />
         ))}
       </div>
@@ -288,10 +297,10 @@ export default function OldItemGame() {
       {correctGuess ? (
         <Victory
           id="victory"
-          championKey={items[0]?.id}
+          championKey={champions[0].key}
           champion={currentGuess}
           tries={guesses.length}
-          isOldItem={true}
+          title={title}
         />
       ) : (
         ""
