+++ frontend/src/views/Splash.jsx
